### 2019-04-14 #1379 @labarba
@IbrahimTanyalcin —You've given us a link to your project webpage, not the repository: it should be https://github.com/IbrahimTanyalcin/lexicon-mono-seq
Let's see if I can change that manually.
