### 2019-04-14 #1379 @labarba
@IbrahimTanyalcin — Before we can start with assigning an editor and reviewers, you'll have to do a bit more work on the paper to meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Right now the article itself is very brief and doesn't contain much substance; while we don't expect a full article here, it should be 250–1000 words and should include:

- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software, and comments on other extant software with similar function

Your article should be improved with a brief description of how the software works, and some examples use cases.
