### 2019-04-15 #1379 @IbrahimTanyalcin
@labarba , I'll beef up the paper.md a bit, the examples are embedded inside the readme, I've got 6 already, 2 more on the way. I'll let you know when I'm done, probably in a few hours.
