### 2019-04-15 #1379 @IbrahimTanyalcin
@labarba, Could you take a quick glance at the paper.md, does it fit better? https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/paper/paper.md
