+++ lexiconMonoSeq.v0.16.1.js
@@ -6,13 +6,13 @@
 		} else {
 			root.LexiconMonoSeq = factory(root,root.document);
 		}
-	}(this, function(window,document) {
+	}(this, function($window,$document) {
 		function LexiconMonoSeq(container,options){
 			container = typeof container === "string" 
-				? document.querySelector(container) 
+				? $document.querySelector(container) 
 				: container.nodeType && container.nodeType === 1
 					? container
-					: document.body;
+					: $document.body;
 			return new _LexiconMonoSeq(container,options);
 		};
 		function _LexiconMonoSeq(container,options){
@@ -20,10 +20,10 @@
 				options && (this.textRendering = options.textRendering);
 				options && (this.fontKerning = options.fontKerning);
 				options && (this.webkitFontSmoothing = options.webkitFontSmoothing);
-				var style = document.createElement("style");
+				var style = $document.createElement("style");
 				style.textContent = this.Style.string(this);
 				style.setAttribute("data-name","LexiconMonoSeqStyle");
-				document.head.appendChild(style);
+				$document.head.appendChild(style);
 				this.constructor.prototype.__defaultStyleApplied = true;
 			}
 			this.container = container;
@@ -34,15 +34,15 @@
 				.map(function(d,i){return new this.Paint(this);},this);
 			painters._counter = 0;
 			painters._length = painters.length;
-			this.wrapper = document.createElement("div");
+			this.wrapper = $document.createElement("div");
 			this.wrapper.className = "LexiconMonoSeq wrapper custom";
-			this.div = this.wrapper.appendChild(document.createElement("div"));
+			this.div = this.wrapper.appendChild($document.createElement("div"));
 			this.div.className = "LexiconMonoSeq main custom";
 			this.divStyle = this.div.style;
-			this.hidden = this.wrapper.appendChild(document.createElement("div"));
+			this.hidden = this.wrapper.appendChild($document.createElement("div"));
 			this.hidden.className = "LexiconMonoSeq hidden";
 			this.hidden.textContent = "A";
-			this.busy = this.wrapper.appendChild(document.createElement("div"));
+			this.busy = this.wrapper.appendChild($document.createElement("div"));
 			this.busy.className = "LexiconMonoSeq busy custom";
 			this.busyStyle = this.busy.style;
 			this.busy.textContent = "\u23f3";
@@ -487,7 +487,7 @@
 			return arr;
 		};
 		_LexiconMonoSeq.prototype.SVG = function(container){
-			var svg = document.createElementNS("http://www.w3.org/2000/svg","svg");
+			var svg = $document.createElementNS("http://www.w3.org/2000/svg","svg");
 			svg.setAttribute("viewBox","0 0 1 1");
 			svg.setAttribute("preserveAspectRatio","none");
 			if (container) {
@@ -572,12 +572,12 @@
 							return this._fontRect;
 						}
 						var that = this,
-							r = document.createRange(),
+							r = $document.createRange(),
 							t = this.hidden.firstChild,
 							l = t.length;
 						r.setStart(t,0);
 						r.setEnd(t,l);
-						window.requestAnimationFrame(function(){
+						$window.requestAnimationFrame(function(){
 							delete that._fontRect;
 							r.detach && r.detach();
 						});
@@ -604,7 +604,7 @@
 							return this._fontRatio;
 						}
 						var that = this;
-						window.requestAnimationFrame(function(){
+						$window.requestAnimationFrame(function(){
 							delete that._fontRatio;
 						});
[…]
