+++ snippets/snippet-1.js
@@ -1,0 +1,171 @@
+
+!function(){
+	var LexiconMonoSeq = require("lexicon-mono-seq");
+	var container = document.createElement("div");
+	container.id = "test";
+	rootDiv.appendChild(container);
+	var	body = document.body,
+		styleText = [
+			'@font-face {font-family: "Consolas";',
+				'src: url("//db.onlinewebfonts.com/t/1db29588408eadbd4406aae9238555eb.eot");',
+				'src: url("//db.onlinewebfonts.com/t/1db29588408eadbd4406aae9238555eb.eot?#iefix") format("embedded-opentype"),',
+				'url("//db.onlinewebfonts.com/t/1db29588408eadbd4406aae9238555eb.woff2") format("woff2"),',
+				'url("//db.onlinewebfonts.com/t/1db29588408eadbd4406aae9238555eb.woff") format("woff"),',
+				'url("//db.onlinewebfonts.com/t/1db29588408eadbd4406aae9238555eb.ttf") format("truetype"),',
+				'url("//db.onlinewebfonts.com/t/1db29588408eadbd4406aae9238555eb.svg#Consolas") format("svg");',
+			'}',
+			'#test {',
+				'width:75vw;',
+				'height:25vh;',
+				'position:absolute;',
+				'margin:auto;',
+				'top:0;',
+				'right:0;',
+				'bottom:0;',
+				'left:0;',
+			'}',
+			'body {',
+				'position:relative;',
+				'margin:0px;',
+				'width:100vw;',
+				'height:100vh;',
+			'}',
+			'#loading {',
+				'display:flex;',
+				'align-items:center;',
+				'justify-content:center;',
+				'width:20vw;',
+				'height:20vh;',
+				'position:absolute;',
+				'top:0;',
+				'right:0;',
+				'bottom:0;',
+				'left:0;',
+				'margin:auto;',
+				'font-size:16px;',
+				'font-family:Consolas;',
+				'text-align:center;',
+				'border-radius:10px;',
+				'background:rgba(0,0,0,0.8);',
+				'color:White;',
+				'z-index:2;',
+			'}',
+			'button {',
+				'width:40vw;',
+				'height:10vh;',
+				'position:absolute;',
+				'top:5vh;',
+				'right:0;',
+				'left:0;',
+				'margin:auto;',
+				'font-size:20px;',
+				'font-family:Consolas;',
+				'text-align:center;',
+				'border:0px;',
+				'padding:0px;',
+				'border-radius:10px;',
+				'background:DodgerBlue;',
+				'color:white;',
+				'outline:none;',
+				'transition:background 1s ease;',
+			'}',
+			'button.inactive {',
+				'background:DarkSlateGray !important;',
+			'}',
+			'button:hover {',
+				'background:Teal;',
+			'}'
+		],
+		style = document.createElement("style"),
+		instance = LexiconMonoSeq(container,{parallelRendering:5}),
+		watchman = instance.constructor.prototype.watchman,
+		testNode = document.createElement("div"),
+		testNodeInitRect,
+		testNodeStyle = testNode.style;
+	testNodeStyle.visibility = "hidden";
+	testNodeStyle.position = "absolute";
+	testNodeStyle.fontSize = "16px";
+	testNodeStyle.fontFamily = "inexistentFont";
+	testNodeStyle.margin = "0px";
+	testNodeStyle.padding = "0px";
+	testNode.textContent = "A";
+	var fontNode = testNode.cloneNode(true),
+		fontNodeInitRect;
+	fontNode.style.fontFamily = "Consolas";
+	body.appendChild(testNode);
+	body.appendChild(fontNode);
+	style.textContent = styleText.join("");
+	document.head.appendChild(style);
[…]
