### 2019-04-17 #1379 @arfon
@IbrahimTanyalcin - this looks pretty good. It looks like you've not cited any of the entries in your `paper.bib` file which means they're not being compiled into the PDF proof. Could you make sure cite these? You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
