### 2019-04-17 #1379 @IbrahimTanyalcin
@arfon ,I'll fix the citations. I've added 2 more examples meanwhile.
