+++ paper/paper.bib
@@ -6,6 +6,47 @@
   YEAR = 2018, 
 } 
 
+@Article{Li2018,
+author={Li, ManZhi
+and Long, HaiXia
+and Wang, HongTao
+and Fu, HaiYan
+and Xu, Dong
+and Shen, YouJian
+and Yao, YuHua
+and Liao, Bo},
+title={Disease Sequences High-Accuracy Alignment Based on the Precision Medicine},
+journal={Biomed Res Int},
+year={2018},
+month={Feb},
+day={22},
+publisher={Hindawi},
+volume={2018},
+pages={1718046-1718046},
+abstract={High-accuracy alignment of sequences with disease information contributes to disease treatment and prevention. The results of multiple sequence alignment depend on the parameters of the objective function, including gap open penalties (GOP), gap extension penalties (GEP), and substitution matrix (SM). Firstly, the theory parameter formulas relating to GOP, GAP, and SM are inferred, combining unaligned sequence length, number, and identity. Secondly, we tested the rationality of the theory parameter formulas, with experiment on the ClustalW and MAFFT program. In addition, we obtained a group of MAFFT program parameters according to the formulas proposed. The results of all experiments show that the SPS (sum-of-pair score) obtained from theory parameters is better than the SPS obtained from the default parameters of ClustalW and MAFFT. In both theory and practice, our method to determine the parameters is feasible and efficient. These can provide high-accuracy alignment results for precision medicine.},
+note={29682519[pmid]},
+note={PMC5842723[pmcid]},
+issn={2314-6141},
+doi={10.1155/2018/1718046},
+url={https://www.ncbi.nlm.nih.gov/pubmed/29682519}
+}
+
+@article{clustal,
+    author = {Wilm, A. and Higgins, D.G. and Valentin, F. and Blackshields, G. and McWilliam, H. and Wallace, I.M. and Thompson, J.D. and Larkin, M.A. and Brown, N.P. and McGettigan, P.A. and Chenna, R. and Lopez, R. and Gibson, T.J.},
+    title = "{Clustal W and Clustal X version 2.0}",
+    journal = {Bioinformatics},
+    volume = {23},
+    number = {21},
+    pages = {2947-2948},
+    year = {2007},
+    month = {09},
+    abstract = "{Summary: The Clustal W and Clustal X multiple sequence alignment programs have been completely rewritten in C++. This will facilitate the further development of the alignment algorithms in the future and has allowed proper porting of the programs to the latest versions of Linux, Macintosh and Windows operating systems.Availability: The programs can be run on-line from the EBI web server: http://www.ebi.ac.uk/tools/clustalw2. The source code and executables for Windows, Linux and Macintosh computers are available from the EBI ftp site ftp://ftp.ebi.ac.uk/pub/software/clustalw2/Contact:clustalw@ucd.ie}",
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btm404},
+    url = {https://doi.org/10.1093/bioinformatics/btm404},
+    eprint = {http://oup.prod.sis.lan/bioinformatics/article-pdf/23/21/2947/1057599/btm404.pdf},
+}
+
 @online{Npm,
   author = {Npm},
   title = {npmjs.com},
