### 2019-04-18 #1379 @IbrahimTanyalcin
@arfon, I've added the refs, though I'm not sure if it'll compile properly.
