### 2019-04-18 #1379 @arfon
I could 6 citations in the `paper.md`, 6 references in `paper.bib` and 6 rendered references in the output PDF. That seems correct?
