### 2019-04-22 #1379 @danielskatz
👋 @csoneson - ping... would you be willing to edit this submission for JOSS?
