### 2019-04-23 #1379 @csoneson
👋 @IbrahimTanyalcin -thanks for your submission, I am in the process of reaching out to possible reviewers. I just had a quick look at your repository, and I could not find a valid license file (see https://joss.readthedocs.io/en/latest/review_criteria.html#software-license) - could you check that?
