### 2019-04-23 #1379 @csoneson
Hi @jkanche - would you be interested in reviewing this submission for JOSS? Looks like a nice fit for your expertise!
