### 2019-04-23 #1379 @jkanche
Hi @csoneson, Yes, I'm interested in reviewing this. This would be my first review for JOSS, I was wondering what the official process is ?
