### 2019-04-23 #1379 @csoneson
@jkanche Awesome! You can find reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Basically, in a few minutes I will assign you as one of the reviewers of the submission, and open the 'actual' review issue (there will be a link posted here). The review will take place in that issue - there will be a checklist for you to go through and evaluate the submission against. Comments and suggestions are also to be posted in that issue.
