### 2019-04-23 #1407 @whedon
[REVIEW]: Lexicon-Mono-Seq, DOM Text Based Async MSA Viewer
**Submitting author:** @IbrahimTanyalcin (<a href="http://orcid.org/0000-0003-0327-5096">Ibrahim  Tanyalcin</a>)
**Repository:** <a href="https://github.com/IbrahimTanyalcin/lexicon-mono-seq" target ="_blank">https://github.com/IbrahimTanyalcin/lexicon-mono-seq</a>
**Version:** v0.19.0
