### 2019-04-23 #1379 @csoneson
Reviewers - please head over to #1407 for the actual review!
