### 2019-04-23 #1407 @csoneson
@labarba (or @openjournals/joss-eics) - just to say that the manual change of the path to the submitted repository in the PRE-REVIEW issue (#1379) did not seem to carry over here. I manually changed it again above, which I think should be fine for the review, but I'm not sure whether anything else needs to be changed somewhere for a more permanent effect.
