### 2019-04-24 #1407 @hrhotz
@IbrahimTanyalcin - I have made a pull request with some minor suggestions for the paper text.
Also, please consider adding the names of a few examples for browsers using DOM Text + SVG. Same for MAS viewers relying on the canvas element. This might help putting the text in a wider context.
