+++ paper/paper.bib
@@ -70,3 +70,76 @@ url={https://www.ncbi.nlm.nih.gov/pubmed/29682519}
   url = {https://github.com/tc39},
   urldate = {2018-02-19}
 }
+
+@Article{SnipViz,
+author={Jaschob, Daniel
+and Davis, Trisha N.
+and Riffle, Michael},
+title={SnipViz: a compact and lightweight web site widget for display and dissemination of multiple versions of gene and protein sequences},
+journal={BMC Res Notes},
+year={2014},
+month={Jul},
+day={23},
+publisher={BioMed Central},
+volume={7},
+pages={468-468},
+abstract={BACKGROUND: As high throughput sequencing continues to grow more commonplace, the need to disseminate the resulting data via web applications continues to grow. Particularly, there is a need to disseminate multiple versions of related gene and protein sequences simultaneously--whether they represent alleles present in a single species, variations of the same gene among different strains, or homologs among separate species. Often this is accomplished by displaying all versions of the sequence at once in a manner that is not intuitive or space-efficient and does not facilitate human understanding of the data. Web-based applications needing to disseminate multiple versions of sequences would benefit from a drop-in module designed to effectively disseminate these data. FINDINGS: SnipViz is a client-side software tool designed to disseminate multiple versions of related gene and protein sequences on web sites. SnipViz has a space-efficient, interactive, and dynamic interface for navigating, analyzing and visualizing sequence data. It is written using standard World Wide Web technologies (HTML, Javascript, and CSS) and is compatible with most web browsers. SnipViz is designed as a modular client-side web component and may be incorporated into virtually any web site and be implemented without any programming. CONCLUSIONS: SnipViz is a drop-in client-side module for web sites designed to efficiently visualize and disseminate gene and protein sequences. SnipViz is open source and is freely available at https://github.com/yeastrc/snipviz.},
+note={25056180[pmid]},
+note={PMC4118779[pmcid]},
+note={1756-0500-7-468[PII]},
+issn={1756-0500},
+doi={10.1186/1756-0500-7-468},
+url={https://www.ncbi.nlm.nih.gov/pubmed/25056180}
+}
+
+@article{MSAViewer,
+    author = {Rauscher, Benedikt and Wilzbach, Sebastian and Goldberg, Tatyana and Rost, Burkhard and Yachdav, Guy and Sheridan, Robert and Sillitoe, Ian and Procter, James and Lewis, Suzanna E.},
+    title = "{MSAViewer: interactive JavaScript visualization of multiple sequence alignments}",
+    journal = {Bioinformatics},
+    volume = {32},
+    number = {22},
+    pages = {3501-3503},
+    year = {2016},
+    month = {07},
[…]
