+++ readme.md
@@ -10,6 +10,8 @@ layout: post
 [![DOI](https://joss.theoj.org/papers/9fbf126b501ad921825f755732450278/status.svg)](https://joss.theoj.org/papers/9fbf126b501ad921825f755732450278)
 ![npm](https://img.shields.io/npm/v/lexicon-mono-seq.svg)
 ![npm](https://img.shields.io/npm/dm/lexicon-mono-seq.svg)
+![GitHub file size in bytes](https://img.shields.io/github/size/IbrahimTanyalcin/lexicon-mono-seq/lexiconMonoSeq.v0.17.3.min.js.svg)
+![jsDelivr hits (npm)](https://img.shields.io/jsdelivr/npm/hm/lexicon-mono-seq.svg?color=orange)
 
 *LexiconMonoSeq* is a multiple sequence alignment viewer ([MSA](https://en.wikipedia.org/wiki/Multiple_sequence_alignment)) that works with *monospace* fonts. You can use this tool for showing any arbitrary number of strings as well. The library renders DOM Text asynchronously.
 
