### 2019-04-24 #1407 @IbrahimTanyalcin
@hrhotz, ty for the suggestions:

- merged your pull request
- added names to the examples
- added jsdelivr/npm link for downloading the script
- added a minified version from jsdelivr (~40kB)
- added 4 new citations, 2 tools that are similar (DOM text) to lexicon-mono-seq, and 2 that are canvas based (IGV and MSA Viewer)
