### 2019-04-26 e3416c3
typos and edits
