+++ paper/paper.md
@@ -3,11 +3,11 @@ title: 'Lexicon-Mono-Seq, DOM Text Based Async MSA Viewer'
 tags:
  - ES5
  - async
- - javascript
+ - JavaScript
  - requestAnimationFrame
  - js
- - biojs
- - bionode
+ - BioJs
+ - Bionode
  - alignment
  - sequence
  - bioinformatics
@@ -25,20 +25,20 @@ bibliography: paper.bib
 
 # Summary
 
-In Bioinformatics, one of the most common problems is depiction of long amino-acid / dna sequences and their comparison [@Li2018]. Routinely, these sequences are aligned and the data is stored in common formats such as [clustal](http://www.clustal.org/download/clustalw_help.txt) [@clustal]. These alignments are called Multiple Sequence Alignments ([MSA](https://en.wikipedia.org/wiki/Multiple_sequence_alignment)) and can be viewed with variety of different [tools](https://en.wikipedia.org/wiki/List_of_alignment_visualization_software).
+In Bioinformatics, one of the most common problems is depiction of long amino-acid / DNA sequences and their comparison [@Li2018]. Routinely, these sequences are aligned to a reference and the data is stored in common formats such as [clustal](http://www.clustal.org/download/clustalw_help.txt) [@clustal]. These alignments are called Multiple Sequence Alignments ([MSA](https://en.wikipedia.org/wiki/Multiple_sequence_alignment)) and can be viewed with variety of different [tools](https://en.wikipedia.org/wiki/List_of_alignment_visualization_software).
 
-*Lexicon-Mono-Seq* is an MSA viewer specially designed for browsers that uses native [DOM Text](https://www.w3.org/TR/dom/#dom-node-textcontent) + [SVG](https://www.w3.org/TR/SVG11/) [@SnipViz,@EnsemblMView] instead of [HTML5 canvas](https://www.w3.org/TR/2011/WD-html5-20110525/the-canvas-element.html) [@MSAViewer,@IGV].
+*Lexicon-Mono-Seq* is an MSA viewer specially designed for browsers that use native [DOM Text](https://www.w3.org/TR/dom/#dom-node-textcontent) + [SVG](https://www.w3.org/TR/SVG11/) [@SnipViz,@EnsemblMView] instead of [HTML5 canvas](https://www.w3.org/TR/2011/WD-html5-20110525/the-canvas-element.html) [@MSAViewer,@IGV].
 
-Several MSA viewers that are already in use rely on rendering characters as an image using the *canvas* element. Although reliable and convenient, this can result in 2 problems: 
+Several MSA viewers rely on rendering characters as an image using the *canvas* element. Although reliable and convenient, this can result in 2 problems: 
   - Higher memory usage,
   - Blurred text (increasing resolution increases memory usage) 
 
-*Lexicon-Mono-Seq* aims to approach alignment viewing problem from the [DOM](https://www.w3.org/DOM/) (Document Object Model) perspective, each character is a combination of a [textNode](https://www.w3.org/TR/DOM-Level-2-Core/core.html#ID-1312295772) and an [SVG Graphics Element](https://developer.mozilla.org/en-US/docs/Web/SVG/Element#Graphics_elements). The script recycles these nodes and paints only regions that are necessary to be painted [@Lexicon]. Comparison of sequences with different character widths are also a possibility, this allows one to compare amino acids and dna regions within the same applet (i.e. one amino acid every three nucleotides in the dna). *Lexicon-Mono-Seq* is agnostic about the character width of sequences, a dataset can have multiple sequences with different character widths.
[…]
