### 2019-04-26 #1407 @jkanche
@IbrahimTanyalcin - I created a pull request with a few minor edits and fixes to the paper. Couple of things

- version number of the release does not match the one in the paper
- I'd consider adding this to examples
   - https://github.com/dariober/ASCIIGenome
- renaming files in examples with specific names so that its easier to browse
