+++ paper/paper.bib
@@ -141,5 +141,19 @@ url={https://doi.org/10.1038/nbt.1754}
 	URL = {http://europepmc.org/articles/PMC5570243},
 }
 
-
+@article{asciigenome,
+    author = {Beraldi, Dario},
+    title = "{ASCIIGenome: a command line genome browser for console terminals}",
+    journal = {Bioinformatics},
+    volume = {33},
+    number = {10},
+    pages = {1568-1569},
+    year = {2017},
+    month = {01},
+    abstract = "{Current genome browsers are designed to work via graphical user interfaces (GUIs), which, however intuitive, are not amenable to operate within console terminals and therefore are difficult to streamline or integrate in scripts. To circumvent these limitations, ASCIIGenome runs exclusively via command line interface to display genomic data directly in a terminal window. By following the same philosophy of UNIX tools, ASCIIGenome aims to be easily integrated with the command line, including batch processing of data, and therefore enables an effective exploration of the data.ASCIIGenome is written in Java. Consequently, it is a cross-platform tool and requires minimal or no installation. Some of the common genomic data types are supported and data access on remote ftp servers is possible. Speed and memory footprint are comparable to or better than those of common genome browsers.Software and source code (MIT License) are available at https://github.com/dariober/ASCIIGenome with detailed documentation at http://asciigenome.readthedocs.io.Supplementary data are available at Bioinformatics online.}",
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btx007},
+    url = {https://doi.org/10.1093/bioinformatics/btx007},
+    eprint = {http://oup.prod.sis.lan/bioinformatics/article-pdf/33/10/1568/25153211/btx007.pdf},
+}
 
