+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 In Bioinformatics, one of the most common problems is depiction of long amino-acid / DNA sequences and their comparison [@Li2018]. Routinely, these sequences are aligned to a reference and the data is stored in common formats such as [clustal](http://www.clustal.org/download/clustalw_help.txt) [@clustal]. These alignments are called Multiple Sequence Alignments ([MSA](https://en.wikipedia.org/wiki/Multiple_sequence_alignment)) and can be viewed with variety of different [tools](https://en.wikipedia.org/wiki/List_of_alignment_visualization_software).
 
-*Lexicon-Mono-Seq* is an MSA viewer specially designed for browsers that use native [DOM Text](https://www.w3.org/TR/dom/#dom-node-textcontent) + [SVG](https://www.w3.org/TR/SVG11/) [@SnipViz,@EnsemblMView] instead of [HTML5 canvas](https://www.w3.org/TR/2011/WD-html5-20110525/the-canvas-element.html) [@MSAViewer,@IGV].
+*Lexicon-Mono-Seq* is an MSA viewer specially designed for browsers that use native [DOM Text](https://www.w3.org/TR/dom/#dom-node-textcontent) + [SVG](https://www.w3.org/TR/SVG11/) [@SnipViz,@EnsemblMView] instead of [HTML5 canvas](https://www.w3.org/TR/2011/WD-html5-20110525/the-canvas-element.html) [@MSAViewer,@IGV] or CLI [@asciigenome].
 
 Several MSA viewers rely on rendering characters as an image using the *canvas* element. Although reliable and convenient, this can result in 2 problems: 
   - Higher memory usage,
@@ -38,7 +38,7 @@ Several MSA viewers rely on rendering characters as an image using the *canvas* 
 At its core *Lexicon-Mono-Seq* is a monospace font ascii character animation framework and comes with its own pre-baked easing functions, much like [CSS Cubic Bezier](https://developer.mozilla.org/en-US/docs/Web/CSS/timing-function) functions, these provide customizable and cached timing points for animations. Other fields such as creative coding or data visualization can use *Lexicon-Mono-Seq* for transitioning between ascii maps or large heatmaps.
 
 With minimal technical debt (no dependencies), the library uses [requestAnimationFrame](https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame) (rAF) to asynchronously render blocks of sequences with the desired animation, duration and node limit.
-The library is about 11kB (as of version 0.15.12) minimized and does not need any transpiling to use in older browsers that do not support ES6 [@TC39].
+The library is about 42kB (as of version 0.17.3) minimized and does not need any transpiling to use in older browsers that do not support ES6 [@TC39].
 The amount of sequences is limited to your browsers capacity, recommended limit is 8000 sequences with 100000 characters long each.
 
 *Lexicon-Mono-Seq* is suitable for developers who only want the functionality of a minimal MSA Viewer but desire to design their own interface around it. The script provides several utility functions that can be hooked to custom buttons as desired [@Npm;@GitHub].
