+++ docs/index.md
@@ -110,7 +110,7 @@ Above will skip 30 frames (considering 60 frames = 1 second) and then update, gi
 
 ![example1](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example1.gif)
 
-> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/example-1.html)
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/1000x10000RandomSeq.html)
 > #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-1.html)
 
 *LexiconMonoSeq* comes with static (directly attached to the main function) methods that allows you to attach a *ruler* to your object, it finds the longest sequence and adds a ruler either to top of bottom of your dataset:
@@ -130,7 +130,7 @@ The second argument above are additional options which you can pass to *LexiconM
 
 ![example2](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example2.gif)
 
-> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/example-2.html)
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/dynamicUpdate.html)
 > #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-2.html)
 
 You do not have to instantiate a new version each time your object changes, you can repeteadly call <code>update</code> method to redraw sequences. Things to be removed/added/changed are arranged automatically.
@@ -139,7 +139,7 @@ You do not have to instantiate a new version each time your object changes, you 
 
 ![example3](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example3.gif)
 
-> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/example-3.html)
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/PI10Million.html)
 > #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-3.html)
 
 If you ever need to get coordinates and data from *LexiconMonoSeq*, you can pass the evet to its <code>getInfoFromEvent</code> method:
@@ -168,7 +168,7 @@ You can also use the <code>getInfoFromRect</code> method to receive sequences fr
 
 ![example4](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example4.gif)
 
-> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/example-4.html)
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/randomSequenceSmoothScroll.html)
 > #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-4.html)
 
 You can automatically scroll to any horizontal/vertical position, for a dataset of 10000 letters length max and 1000 sequences:
@@ -182,7 +182,7 @@ Above would scroll to some random position along the map.
 
 ![example5](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example5.gif)
 
-> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/example-5.html)
[…]

[…]
