### 2019-04-26 #1407 @IbrahimTanyalcin
@jkanche, ty for the suggestions
- merged your PR
- add a new reference for ascii genome
- changed the example names in the repo
