### 2019-04-29 #1407 @hrhotz
@IbrahimTanyalcin - I have been looking at the "clustalW.html" example and I have a few questions/comments:

- what is the "aa" color scheme based on?
- would it make sense to offer more than just one color set for amino acids?
- how do you deal with sequences containing Selenocysteine? I have modified your example by replacing an "A" with and "U" and the complete sequence is no longer displayed,
- why is the line describing the identity/similarity (the line containing "*",":", etc) not displayed
- I don't understand your comment: "You can also read clustal files if you provide them to LexiconMonoSeq as a string". You already provide the data as a string in the example, don't-you?
- displaying an alignment where you provide a file path would be useful.
