### 2019-04-29 #1407 @imallona
Hi @IbrahimTanyalcin,

Thanks for the resource. I have some comments/questions regarding the rendered PDF version of the paper, and the resource and its documentation.

I noticed the paper update has fixed some typos already (dna -> DNA), but I still find some in citations. Namely,
- ES6 citation is currently is listed as (International, 2018) instead of (Ecma International, 2018).
- The reference for lexicon library by _Tanyalcin et al._ does not indicate the journal name.
- Also, this paper footpage `Tanyalcin, (2019).` doesn't list the author's full name, but only the last name.

Regarding the tool usability, it is stated that _(...) the amount of sequences is limited to your browsers capacity, recommended limit is 8000 sequences with 100000 characters long each (...)_. Is it possible to tabulate/stratify such recommendation, i.e. providing a limit for commodity laptops as well as to recent/powerful machines/workstations?

Regarding the tool usability and/or documentation, I acknowledge the high flexibility given to the `durations` of animation effects, which (if I'm not wrong) can be quite short, i.e. `17ms`. As intermittent visual stimuli are known to elicit epileptic seizures to some, I wonder which are your thoughts on including a warning at the resource documentation suggesting good practices. The W3C indicates that, in order to avoid triggering seizures to photosensitive epilepsy users, blinks should be slower than three a second https://www.w3.org/TR/2008/REC-WCAG20-20081211/#seizure . Not sure of the examples' animation speeds (the ones rendered at `readme.md`), but if they were faster than that I would suggest slowing them down to the W3C recommendations.
