### 2019-04-30 #1407 @IbrahimTanyalcin
@hrhotz  Thank you for the well raised comments, let me go over some of these:

- the "aa" (amino acid) color scheme is from a thread I found here: http://ugene.net/forum/YaBB.pl?num=1337064665    not perfect but made some sense at the time.
- You can add any custom color scheme, using the [`instance.registerType`](https://github.com/IbrahimTanyalcin/lexicon-mono-seq#methods) method. It looks like this:

```
instance.registerType("myType",{"A":"#ffbbaa","B":"00aacc"...})
```
You can also specify a text color and opacity which are optional, default text color is [rgb(0,0,0,0.9)](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L1412) and default opacity for any character is [0.8](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L1548)

Normally you do not have to create a new object from scratch to pass to `instance.registerType` as an argument, a copy function exists [`instance.caveManCopy`](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L294) on the prototype chain. Example-7(chrX10Mb.html in the repo examples folder) makes use of [this](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/chrX10Mb.html#L99-L109). It registers a new type with new characters.

The error you encountered (things not rendering) is possibly because the new char you specified was not registered in the type and it threw `Uncaught Reference Error`, I am not using try/catch blocks there because I presume it would slow down execution.

- For the similarity line with "*" character for Clustal.wl files, I initially thought the coloring would be sufficient to draw human perception to similar regions, but now after your comment I realize it might not be adequate. The static function that reads the clustal.wl is [here](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L1124) and specifically [this is the line](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L1128) that "discards" the lines with "*". I propose that we pass an optional argument to not filter out those lines. What do you think.

- That is correct, I will revise that sentence, "also" is unnecessary. Thank you for pointing out.
- Indeed, currently in the examples I use XMLHttp (not fetch for older browsers). I propose I add a static "fetch" function to read out clustal.wl files. An options arguments in the form of an object could be passed where keys:
    - retry : specify retry count upon failure
    - medium: either "local" (for file API) or "remote" (default, uses XMLHttp)
   The above would return a thenable object which you can chain later on. Is that ok?
