### 2019-04-30 #1407 @IbrahimTanyalcin
@imallona Thank you for pointing out these blind spots. I will go over the citations, did not realize that they were incomplete.

For the performance, I just tested the [first example](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/1000x10000RandomSeq.html) and the [letters of PI](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/PI10Million.html) on an old iphone 5S to see if it did render. The animations were not butter smooth, however it handled without problems (iphone 8 and X rendered without problems and with 60fps). Since we are working with DOM, we are mainly using CPU (apart from transform css style, which creates a layer for GPU) and this could be quite variable. That being said, since sequence length is not the bottle neck and to be on the safe side, we could say for most commodity laptops 1000 sequences with 100000 letters each could run reasonably smooth (since they run on iphone 5S). I have access to variety of different devices including android, iOS and Mac. If you have specific hardware let me know so I can test it on them.

About the epileptic seizures, thank you for raising this point! I must say I am completely ignorant in this field and I think it is a good idea to at least include some sort of disclaimer. For the duration, you are correct that the lowest you can go is 17ms (this is hardcoded in [`window.requestAnimationFrame`](https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame) API). The default duration is [1500milliseconds](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L2032), this controls the sliding of the text. The colored boxes are filled with a default duration of [150ms](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L1811) with a limit of [80 characters per line](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L1810). Another parameter is also the painter count (these are parallel async renderers), by default there is [only 1 painter](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.17.3.js#L33), an example where you can spawn more painters is [here](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/clustalW.html#L46).

What above means is that, by default if a display has 15 sequences with a screen with of 120 characters, it would take Math.ceil(120/80) * 15 * 150 + (15-1) * 17 ~= 4750ms to fill the screen and the screen would briefly "update" 30 times within this period, that is roughly 6 "updates" per second which is twice the recommended range in the link you kindly provided.

I suggest we add a disclaimer, and for some [examples like this](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/PI10Million.html#L130), we could increase the duration to 350ms. What do you think?
