### 2019-04-30 #1407 @hrhotz
@IbrahimTanyalcin - Thanks for your feedback, please allow me to go a little bit deeper into some of my comments (which I now have numbered) wrt to the "clustalW.html" example:

1. - your chosen "aa" color scheme does make sense. Though, you should probably acknowledge the person who set it up originally.
2. - the option to add a custom color scheme is great. It might help if you document this (with a code example) in addition to the code example shown for the "asciiArt" and/or add the text you have written for "Example-7(chrX10Mb.html)" in this thread to the documentation. However, my actual comment was: whether it would make sense to have several pre-defined color schemes which you can chose from. As an example have a look at Jalview (http://www.jalview.org/)
3. - having a "U" in the sequence should not threw an 'Uncaught Reference Error'. Nowadays, "U" (aka Selenocysteine) is very common in amino acid sequences (e.g.: https://www.uniprot.org/uniprot/P07203#sequences ).
4. - an optional argument sound very good
5. -
6. - sure, whatever is the easiest way to implement it, is fine
