### 2019-05-02 #1407 @imallona
Hi @IbrahimTanyalcin,

Thank you for your feedback. I see, so setting thresholds for performance is a difficult task due to the huge variety of hardware available. My comment arose from the paper's sentence _The amount of sequences is limited to your browsers capacity, recommended limit is 8000 sequences with 100000 characters long each_, which I found not very useful as it does not specify what hardware that recommendation applies to (as you say, CPU, operating system/browser). I think that, if updated to add the recommendations for commodity laptops/phones you kindly posted here (1000 sequences with 100000 letters each), that sentence would be more informative, what do you think?

As for the update rate I appreciate your reflection and agree with your suggestions of adding a disclaimer and slowing down some examples.

Aside, as a minor comment unrelated to the former, I just noticed the `paper.md` refers to version `0.17.3` _(...)the library is about 42kB (as of version 0.17.3)(...)_; you might want to update both library version and its size. Also noticed that `dna` and `ascii` should be capitalized.
