### 2019-05-15 #1407 @csoneson
👋 @jkanche - just checking in on your review status. There are several unticked boxes in your checklist - could you provide some comments for @IbrahimTanyalcin explaining your current reservations in some more detail?
