### 2019-05-18 #1407 @jkanche
@csoneson, I did send a pull request with suggestions and updates to the paper, and I was waiting to hear back from @IbrahimTanyalcin until he's ready for the review again since he said he's working on new updates and integration in his last comment
