### 2019-05-27 #1407 @IbrahimTanyalcin
:wave:  @csoneson - I finished almost adding it to MutaFrame. I was thinking of adding Web Component based constructor to it, however it might be far fetched. I'll only push what has been requested by tomorrow:
- Selenocysteine
- Support for clustal.wl
- Doc updates
I'll send a notification when I'm finished
