+++ lexiconMonoSeq.v0.18.0.js
@@ -107,7 +107,7 @@
 		##############PROTO############
 		###############################
 		*/
-		_LexiconMonoSeq.prototype.version = "0.17.3";
+		_LexiconMonoSeq.prototype.version = "0.18.0";
 		_LexiconMonoSeq.prototype._scrollLeftOffset = 0;
 		_LexiconMonoSeq.prototype._50spaces = "                                                  ";
 		_LexiconMonoSeq.prototype.set = function(k,v){
@@ -119,10 +119,54 @@
 		_LexiconMonoSeq.prototype.generateRandomString = function(){
 			return (Math.random()*1e9 | 0).toString(16);
 		};
+		_LexiconMonoSeq.prototype.common = {
+			opacities: {
+				A: 0.95,
+				B: 0.95,
+				C: 0.95,
+				U: 0.95,
+				D: 0.95,
+				E: 0.95,
+				F: 0.95,
+				G: 0.95,
+				H: 0.95,
+				I: 0.95,
+				K: 0.95,
+				L: 0.95,
+				M: 0.95,
+				N: 0.95,
+				P: 0.95,
+				Q: 0.95,
+				R: 0.95,
+				S: 0.95,
+				T: 0.95,
+				V: 0.95,
+				W: 0.95,
+				X: 0.95,
+				Y: 0.95,
+				Z: 0.95,
+				"?": 0.8,
+				"-": 0.8,
+				":": 0.4,
+				"*": 0.2,
+				".": 0.1,
+				"·": 0.05,
+				" ": 0
+			},
+			opacitiesPunctuation: {
+				"-": 0.8,
+				":": 0.4,
+				"*": 0.2,
+				".": 0.1,
+				"·": 0.05,
+				" ": 0
+			}
+		};
 		_LexiconMonoSeq.prototype.colors = {
 			aa: {
 				"A": "#ffffc9",
 				"C": "#e3f9ad",
+				"U": "#e3f9ad",
 				"E": "#f93333",
 				"D": "#fb7979",
 				"G": "#c0c0c0",
@@ -141,6 +185,326 @@
 				"W": "#85b0cd",
 				"V": "#ffffab",
 				"Y": "#7dafb9",
+				"X": "#333333",
+				"?": "#333333",
+				"-": "#778899",
+				":": "#778899",
+				"*": "#778899",
+				".": "#778899",
+				"·": "#778899",
+				" ": "#778899"
+			},
+			jalviewClustal: {
+				A: "#80a0f0",
+				B: "#ffffff",
+				C: "#f08080",
+				U: "#f08080",
+				D: "#c048c0",
+				E: "#c048c0",
+				F: "#80a0f0",
+				G: "#f09048",
+				H: "#15a4a4",
+				I: "#80a0f0",
+				K: "#f01505",
+				L: "#80a0f0",
+				M: "#80a0f0",
+				N: "#00ff00",
+				P: "#ffff00",
+				Q: "#00ff00",
+				R: "#f01505",
+				S: "#00ff00",
+				T: "#00ff00",
+				V: "#80a0f0",
+				W: "#80a0f0",
+				X: "#ffffff",
+				Y: "#15a4a4",
+				Z: "#ffffff",
+				"?": "#333333",
+				"-": "#778899",
+				":": "#778899",
+				"*": "#778899",
+				".": "#778899",
+				"·": "#778899",
+				" ": "#778899"
+			},
+			jalviewZappo: {
+				A: "#ffafaf",
+				B: "#ffffff",
+				C: "#ffff00",
+				U: "#ffff00",
+				D: "#ff0000",
+				E: "#ff0000",
+				F: "#ffc800",
[…]
