+++ docs/index.md
@@ -233,6 +233,98 @@ The above example displays the first million bases of chromosomeX with 3 possibl
 > #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-8.html)
 > #### [Test @Observable](https://observablehq.com/@ibrahimtanyalcin/ascii-warp-github)
 
+### 9 Reading Clustal With Conservation
+
+You can also opt to show the conservation in Clustal WL files:
+
+```JavaScript
+LexiconMonoSeq.readClustal(
+    "CLUSTAL W (1.82) multiple sequence alignment....",
+    {conservation:true}
+);
+```
+
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/clustalW_2.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/clustalW_2.html)
+
+### 10 Choosing color schemes
+
+There are different color schemes you can choose from, the classic "aa" scheme is based on http://ugene.net/forum/YaBB.pl?num=1337064665, the jalview based schemes are taken from http://www.jalview.org/help/html/colourSchemes/index.html
+
+```JavaScript
+- aa //amino acid
+- jalviewClustal 
+- jalviewZappo
+- jalviewTaylor
+- jalviewHydrophobicity
+- jalviewHelixPropensity
+- jalviewStrandPropensity
+- jalviewTurnPropensity
+- jalviewBuriedIndex
+- jalviewNucleotide
+- jalviewPurinePyrimidine
+- dna //DNA
+- ruler //used in sequence numbering
+- alphabet //generic letters
+- number
+```
+
+You can always register new colors using `registerType` method:
+```JavaScript
+instance.registerType(
+    "myType", //type name
+    {"A":"#ffbbaa","B":"00aacc"...}, //colors for each letter box
+    "rgba(0,0,0,0.9)", //color for text
+    {"A":0.95, "B": 0.8...} //opacity for each letter box
+);
+```
+
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/colorSchemes.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/colorSchemes.html)
+
+### 11 Fetching Data
+
+LexiconMonoSeq has a built-in fetch mechanism which returns an async operator much like `Promises` in ES6:
+
+```JavaScript
+instance
+.fetch("//someurl.com/lexicon-mono-seq/datasets/file.json","json")
+.then(function(json){
+    //do something with the json
+    //example: this.update(json)
+})
+```
+In certain situations you might have to call fetch in an already async operator, in that case to avoid too much callbacks, you can do:
+
+```JavaScript
+instance
+.async(function(){
+    var response = {value:null,done:false};
+    setTimeout(function(){
+        response.value = "//someurl.com/lexicon-mono-seq/datasets/file.json";
+        response.done = true;
+    },1000);
+    return response;
+}).then(function(retValue){
+    var response = {value:null,done:false};
+    this
+    .fetch(retValue,"json")
+    .then(function(json){
+        response.value = json;
+        response.done = true;
+    });
+    return response
+}).then(function(retValue){
+    //do something with the return value (retValue);
+    //inside thenable, this refers to the instance
+})
+```
+
+Since `fetch` returns an async thenable, you can always call `catch` method for catching errors.
+
+> #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/asyncFetch.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/asyncFetch.html)
+
 ## Creating an instance
 
[…]

[…]
