+++ docs/index.md
@@ -121,7 +121,7 @@ Second one will wait until the font is loaded and then render. But if the loadin
 ![example1](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example1.gif)
 
 > #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/1000x10000RandomSeq.html)
-> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-1.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/1000x10000RandomSeq.html)
 
 *LexiconMonoSeq* comes with static (directly attached to the main function) methods that allows you to attach a *ruler* to your object, it finds the longest sequence and adds a ruler either to top of bottom of your dataset:
 
@@ -141,7 +141,7 @@ The second argument above are additional options which you can pass to *LexiconM
 ![example2](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example2.gif)
 
 > #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/dynamicUpdate.html)
-> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-2.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/dynamicUpdate.html)
 
 You do not have to instantiate a new version each time your object changes, you can repeteadly call <code>update</code> method to redraw sequences. Things to be removed/added/changed are arranged automatically.
 
@@ -150,7 +150,7 @@ You do not have to instantiate a new version each time your object changes, you 
 ![example3](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example3.gif)
 
 > #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/PI10Million.html)
-> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-3.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/PI10Million.html)
 
 If you ever need to get coordinates and data from *LexiconMonoSeq*, you can pass the evet to its <code>getInfoFromEvent</code> method:
 
@@ -179,7 +179,7 @@ You can also use the <code>getInfoFromRect</code> method to receive sequences fr
 ![example4](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example4.gif)
 
 > #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/randomSequenceSmoothScroll.html)
-> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-4.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/randomSequenceSmoothScroll.html)
 
 You can automatically scroll to any horizontal/vertical position, for a dataset of 10000 letters length max and 1000 sequences:
 
@@ -193,7 +193,7 @@ Above would scroll to some random position along the map.
 ![example5](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example5.gif)
 
 > #### [See script](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/asciiAnimation.html)
-> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/example-5.html)
+> #### [Run Example](https://distreau.com/lexicon-mono-seq/examples/asciiAnimation.html)
[…]
