+++ paper/paper.bib
@@ -1,7 +1,8 @@
 @article{Lexicon, 
-  author = "Tanyalcin, I. and Al Assaf, C. and Ferté, J. and Ancien, F. and Khan, T. and Smits, G. and Rooman, M. and Vranken, W.", 
+  author = "Tanyalcin, Ibrahim and Al Assaf, C. and Ferté, J. and Ancien, F. and Khan, T. and Smits, G. and Rooman, M. and Vranken, W.", 
   Title = {Lexicon Visualization Library and JavaScript for Scientific Data Visualization}, 
-  publisher = "Computing in Science and Engineering", 
+  publisher = "IEEE Computing in Science and Engineering", 
+  journal={IEEE Computing in Science and Engineering},
   doi = {10.1109/MCSE.2018.011111125},
   YEAR = 2018, 
 } 
