### 2019-06-02 #1407 @IbrahimTanyalcin
:wave:  @csoneson @hrhotz  @jkanche @imallona

I hope you all are doing great,
I finally could bring some updates:

- Added disclaimer for seizures https://github.com/IbrahimTanyalcin/lexicon-mono-seq#disclaimer
- Updated paper.md, I might have missed a couple of points
- Added Async operations, these are very similar to ES6 Promises, but instead they are written in ES5 and do not require polyfills https://github.com/IbrahimTanyalcin/lexicon-mono-seq#async-operations
- Added tutorial for registering new color schemes/types
https://github.com/IbrahimTanyalcin/lexicon-mono-seq#registering-new-colors
- Added ~10 color schemes from Jalview https://github.com/IbrahimTanyalcin/lexicon-mono-seq#10-choosing-color-schemes
- Most color schemes now include the letter "U" https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L169
- Added 3 new examples:
    - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#9-reading-clustal-with-conservation
    - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#10-choosing-color-schemes
    - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#11-fetching-data
- A new static method parseURL added, internally used by fetch https://github.com/IbrahimTanyalcin/lexicon-mono-seq#static-methods
- A new async operator `ìsFontLoaded` is added, checks for the users whether the font is loaded, then returns a thenable https://github.com/IbrahimTanyalcin/lexicon-mono-seq#a-word-about-the-fonts
