### 2019-06-05 #1407 @hrhotz
@IbrahimTanyalcin - I am still looking at the "clustalW.html" example referring to the same points as in my comments from April 29/30:

1.         solved
2.         it is great, you have added the other color schemes. But it is not clear to me, how I change from  "aa" to "jalviewClustal " in this example
3.         solved
4.         after adding ",{conservation:true}" it displays the conservation. However, in the example, there are 'tab' instead of blanc spaces, and therefore the line is not displayed correctly.
5.        solved
6.        excuse my ignorance, but it is not clear to me how I can replace the hardcoded clustalW alignment with a link to an alignment file.
