### 2019-06-05 #1407 @hrhotz
@csoneson WRT "General checks": I have added the "OK" for 'Repository' and 'Version'...though, I am still confused about the link for "repository url" and I am not sure about the version: is there any update mechanism in order to change it to 18.0
