### 2019-06-05 #1407 @IbrahimTanyalcin
@hrhotz  No worries, I will write a small section to explain how links can be used. The ClustalW normally specifies a single whitespace, the second example is more recent (https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/examples/asyncFetch.html), I'll try  if I can update the converter to except tabs.
I'll also add a section of how you can change it from "aa" to other color scheme from a String input.
@hrhotz, many thanks for pointing these blind spots!
