### 2019-06-10 #1407 @imallona
Hi @IbrahimTanyalcin, thanks for adding the disclaimer; let me check the compiled version of the updated `paper.md`. Minor comment on versioning: noticed the paper refers to `v0.18` (which is the latest `tag`) but this issue refers to an old `v0.15.12`, which is not the latest release.
