### 2019-06-10 #1407 @csoneson
@hrhotz, @imallona - the version in this issue is set at submission time. When a submission is eventually accepted, the version at that time is archived and a DOI is assigned. So at this point, it is not a problem that the version given here differs from the one in the paper.
@hrhotz - for the repository link, this is ok too. The original link was incorrect at submission; this was later fixed, but the link in the checklist is still the submitted one.
