### 2019-06-12 #1407 @imallona
Thanks @csoneson for the explanation on versioning, and for compiling the proof.

@IbrahimTanyalcin could you please fix the _International, E_ citation to be rendered as _Ecma International_? I also noticed that the _Github, 2018_ reference supporting _(...) utility functions that can be hooked to custom buttons as desired_ points to github.com instead of linking a repository, is this correct?
