### 2019-06-14 #1407 @jkanche
I have a few comments on the paper.

> Several MSA viewers rely on rendering characters as an image using the canvas element.
> Although reliable and convenient, this can result in 2 problems: - Higher memory usage, -
> Blurred text (increasing resolution increases memory usage)

The blurred text is probably because of a HiDPI display. you probably have to calculate the pixel density and then create a canvas with larger resolution, but yes, higher resolution uses higher memory.

> The amount of sequences is limited to your browsers capacity, recommended limit is 1000 sequences with 100000 letters each for commodity laptops/mobile devices.

Browsers can show any amount of text. I think the limitation here is with the number of SVG elements you can add to the DOM. This is why canvas is preferred over d3 for large data sets (a single canvas element vs thousands of SVG nodes). One advantage you have here is there's no interactivity with your SVG elements. Also was this number (1000 sequences, each 100000 bp) based on some benchmark testing you did with the library ?

I don't understand the GitHub and NPM citations at the end of the article.

I was thinking, moving the clustal examples to the top might be better.
