### 2019-06-14 #1407 @IbrahimTanyalcin
:wave:  @csoneson @hrhotz @imallona @jkanche

I think I'll be able to do corrections and the section I promised to @hrhotz over the weekend. Just to quickly answer @jkanche,

> Browsers can show any amount text

That I guess for certain limits, that seems correct. However let's put that to test:

https://stuff.mit.edu/afs/sipb/contrib/pi/pi-billion.txt

This loads about a billion letters of pi, as the page grows you'll notice scroll performance will deteriorate.
This means that even a single DOM phasing element with a single textNode has a cost.
For that reason I have taken this approach:

- Whatever not visible on the current dimensions of the [viewport don't render](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L922-L1179),
- If previous textNodes exist, [update them](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L2321-L2330)

> I think the limitation here is with the number of SVG elements you can add to the DOM

This is also not completely correct, for instance here:
> https://distreau.com/lexicon-mono-seq/examples/PI10Million.html
- We never have 10 million svg rects here. They are only added as much as we need.
- Moreover if they are removed, they are NOT garbage collected, they are saved [here](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L2079-L2142) and are ultimately transferred in bulk with [document fragments](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L2021-L2028)
- On top of that the rendering process is completely async, and it is rendered in pieces([here ](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L2305-L2310)and [here](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L33)). If scrolling occurs before the scene is done, the elements are recycled and repositioned.
- The animations are not calculated dynamically, [they are cached, and you can pretty much define any sort of function using 2 control points of cubic bezier](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L2149-L2205)

So when above combined, if traditional canvas without GPU is used, this approach (DOMText + svg) will be mostly faster and have lighter memory footprint. I did several tests my self, you can also check the performance from chrome devtools.

Although current approach has to use CPU mainly, in real life comparison with state of the art MSA tools that do use GPU, I have not seen performance differences.

I did the 1000x10000 benchmark on iphone5, it was good enough so I said let's keep it there. For desktop it can run up to 8000, gets slower afterwards because I have animations.

> One advantage you have here is there's no interactivity with your SVG elements

You can interact with them, there are functions to get [their relative positions here](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L1191-L1283), these functions return the entire sequence [ledger with DOM nodes](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/lexiconMonoSeq.v0.18.0.js#L1230), so you can pretty much do anything with them without impacting performance
