### 2019-06-14 #1407 @jkanche
> That I guess for certain limits, that seems correct. However let's put that to test:
>
> https://stuff.mit.edu/afs/sipb/contrib/pi/pi-billion.txt

The issue here is not with the text, its with the transfer time, since the browser is rendering the text and streaming chunks of the text file at the same time. This causes the lag, but once everything is in memory, the page loads the entire text pretty quickly. for me, It took 5.6 minutes (from devtools) to transfer the entire file from the server to the client.

> The amount of sequences is limited to your browsers capacity, recommended limit is 1000 sequences with 100000 letters each for commodity laptops/mobile devices.

> I think the limitation here is with the number of SVG elements you can add to the DOM

My comment was based on the statement in the paper. So if the page never renders the entire sequence(s), I think the paper doesn't have to mention this as a limitation ?

> if traditional canvas without GPU is used, this approach (DOMText + svg) will be mostly faster and have lighter memory footprint. I did several tests my self, you can also check the performance from chrome devtools.

were these tests documented anywhere in the GitHub repository ?

I didn't know that you can add interactivity to the elements, thanks for clarifying that up.
