### 2019-06-15 #1407 @IbrahimTanyalcin
@jkanche,

- The limitation 1000x10000 is a good idea due the ledgers (these are js objects) that keep track of each sequence row. The string length per row is not that important, you can have 1x10^6 letters or so (if I remember correctly string operations under the hood was reliant on C++ on chrome, so they are fast.) But the amount of track count starts to tire out the CPU.

- no I have not added the profile files from chrome. Jsdelivr has 50mb package limit. However I can upload them on one of my domains so people can pull out if they need to and give then links on the repo. Do you have any preference for software to compare with? I'll do them tomorrow
