+++ paper/paper.bib
@@ -65,7 +65,7 @@ url={https://www.ncbi.nlm.nih.gov/pubmed/29682519}
 }
 
 @online{TC39,
-  author = {Ecma International},
+  author = {{\relax Ecma International}},
   title = {Ecma TC39},
   year = 2018,
   url = {https://github.com/tc39},
