+++ docs/index.md
@@ -37,7 +37,6 @@ In truth this small script is a ascii animation framework that comes with perfor
 ![example8](https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/images/example8.gif)
 
 
-
 ## Usage
 ```JavaScript
 LexiconMonoSeq(
@@ -58,69 +57,33 @@ LexiconMonoSeq(
     }
 ]);
 ```
-## Loading the script
-
-*LexiconMonoSeq* supports [*UMD*](https://github.com/umdjs/umd). Use a derivate of the below link, based on the version you want:
-
-`<script src="https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq/lexiconMonoSeq.v0.16.0.js"></script>`
-
-Above link can change if a new release is published. You can try adding the repo version (0.16.0 etc.):
-
-`<script src="https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq@[version]/lexiconMonoSeq.v[version].js"></script>`
-
-For example:
-
-> Npm
-
-`<script src="https://cdn.jsdelivr.net/npm/lexicon-mono-seq@1.2.5/lexiconMonoSeq.v0.17.3.js"></script>`
-
-> Github
-
-`<script src="https://cdn.jsdelivr.net/gh/IbrahimTanyalcin/lexicon-mono-seq@0.16.0/lexiconMonoSeq.v0.16.0.js"></script>`
-
-Alternatively you can requests all versions from here:
-
-`<script src="https://distreau.com/lexicon-mono-seq/js/lexiconMonoSeq.v0.16.0.js"></script>`
-
-Include the tag either within `body` or `head`, synchronous or asynchronous, according to your page's desing needs.
-
-## A word about the fonts
-
-> **Note:** As of version 0.18.0, I highly recommend using `isFontLoaded` method 
-
-*LexiconMonoSeq* works with [monospace fonts](https://en.wikipedia.org/wiki/Monospaced_font) only, since fixed width characters are desired for proper alignment. Nonetheless, due to vender differences in font rendering, a few millipixel shift is normal.
-
-You can use your custom monospace fonts, but make sure they are properly loaded before you create an instance. If the fonts are loaded later or you changed them, invoke the `reDraw` method of the instace to recalculate styles:
-
-```JavaScript
-instace.reDraw();
-```
-
-If you really wanna be on the safe side for both Mac and Windows, use *Courier New* font. 
-
-If you load your custom fonts with `@import`, `@font-face` or `link` etc., give a small break after you create an instance, to allow browser calculations to kick-in. This time is possibly vendor dependent and usually a tick (~17ms) is sufficient. So instead of:
-
-```JavaScript
-LexiconMonoSeq("#test",{parallelRendering:5}).update(yourDataSet,{durationPaint:500,duration:500})
-```
-Do this:
-
-```JavaScript
-LexiconMonoSeq("#test",{parallelRendering:5}).skipFrames(30).then(function(){this.update(yourDataSet,{durationPaint:500,duration:500})}).then(...
-```
-Or even better:
-
-```JavaScript
-LexiconMonoSeq("#test",{parallelRendering:5}).isFontLoaded(200,2000).then(function(){this.update(yourDataSet,{durationPaint:500,duration:500})}).then(...
-```
-
-First one will skip 30 frames (considering 60 frames = 1 second) and then update, giving browser to correctly render the custom font.
-
-Second one will wait until the font is loaded and then render. But if the loading takes more than 2000ms (2 seconds), it will return an error instead.
 
 ## Examples
 
-### 1 1000 X 10000 Random Sequences
+### 1 Reading Clustal.wl Files
+
[…]
