### 2019-06-17 #1407 @IbrahimTanyalcin
wave @csoneson @hrhotz @imallona @jkanche

Here are some updates:

- The npm and github references are intentional yes, I thought I'd give credit to them for all the hosting etc.
- The Ecma International should be in full now.
- A new version 0.19.0 is released, the async operators are almost identical to how Promises work.
- Exercise order changed with suggestions of @jkanche:
   - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#examples
- Async operator section is updated, and how both observer or promise pattern can be used is explained:
  - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#async-operations
- With suggestions of @hrhotz, reading Clustal and JSON sections are added:
    - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#reading-remote-clustal-files
    - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#reading-json-files
- I have forgotten to add posOffset:
   - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#updating-data
- With suggestions of @jkanche, a performance section has been added, seems like lexicon is a bit faster with better load times, however I tried to emphasize that this is not about performance, it is about providing a more familiar DOM text based alternative. The profiles are downloadable in an indexed location listed at below link:
   - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#performance
- If you want to see the tests live, I have 2 fiddles, one for lexicon-mono-seq and one for the other, you need to wait a bit for the MSAViewer:
   - lexicon-mono-seq:  https://jsfiddle.net/ibowankenobi/apjcxq1w/
   - MSAViewer: https://jsfiddle.net/ibowankenobi/s97g4vxa/
- A remedies for bugs and os specific glitches has been added:
   - https://github.com/IbrahimTanyalcin/lexicon-mono-seq#remedies
