### 2019-06-29 #1407 @csoneson
@IbrahimTanyalcin Looking at the remaining points in the reviewers' checklists above, could you add a couple of statements to the README with guidelines for people who may want to contribute/seek help, and pointers to how to best test the functionality of the software (including how to verify that the output is correct)? There is one reference (Robinson et al) where the `https://doi.org/` part of the DOI should be removed.
@hrhotz, @imallona, @jkanche - please comment if you have additional comments, or let me know (and fill the checklist) if you are happy with the current state. Thanks!
