### 2019-06-30 #1407 @imallona
Hi @csoneson and @IbrahimTanyalcin, I'm happy with the current version, thanks.
