### 2019-06-30 #1407 @IbrahimTanyalcin
:wave:  @csoneson
- added contributing section to readme
- added eslint and js-beautify templates if needed for PRs
- added a doi field to Robinson et. al.
