### 2019-07-01 #1407 @hrhotz
I have just ticked the last box and I am happy to recommend the current version for acceptance.
@IbrahimTanyalcin it was a pleasure to work with you
@imallona and @jkanche thanks for the collaboration
@csoneson thank you very much for the excellent guidance
