### 2019-07-01 #1407 @IbrahimTanyalcin
@hrhotz , thank you for the great input for the color schemes and url fetching idea. The pleasure is mine.
Also, @imanolla thank you for pointing out very important points including w3c recommendation for animation fps, which I had no idea about.
I don't know if @jkanche  is here, but I also thank him for performance related issues.
@csoneson, thank you for checking if all is on the track.
I'm sorry to bother all of you during summer time for the review. I really appreciate that.
