### 2019-07-06 #1407 @csoneson
👋 @jkanche - could you let us know whether you have additional comments for @IbrahimTanyalcin or if you are happy with the current state of the submission? Thanks!
