### 2019-07-08 ba62a9a
Small edits to paper+bibliography
