+++ paper/paper.bib
@@ -1,9 +1,12 @@
 @article{Lexicon, 
 	author = "Tanyalcin, Ibrahim and Al Assaf, C. and Ferté, J. and Ancien, F. and Khan, T. and Smits, G. and Rooman, M. and Vranken, W.", 
-	Title = {Lexicon Visualization Library and JavaScript for Scientific Data Visualization}, 
+	Title = {Lexicon Visualization Library and {JavaScript} for Scientific Data Visualization}, 
 	publisher = "IEEE Computing in Science and Engineering", 
 	journal={IEEE Computing in Science and Engineering},
 	doi = {10.1109/MCSE.2018.011111125},
+	volume = {20},
+	issue = {1},
+	pages = {50-65},
 	YEAR = 2018, 
 } 
 
@@ -17,13 +20,13 @@
 	and Yao, YuHua
 	and Liao, Bo},
 	title={Disease Sequences High-Accuracy Alignment Based on the Precision Medicine},
-	journal={Biomed Res Int},
+	journal={Biomedical Research International},
 	year={2018},
 	month={Feb},
 	day={22},
 	publisher={Hindawi},
 	volume={2018},
-	pages={1718046-1718046},
+	pages={1718046},
 	abstract={High-accuracy alignment of sequences with disease information contributes to disease treatment and prevention. The results of multiple sequence alignment depend on the parameters of the objective function, including gap open penalties (GOP), gap extension penalties (GEP), and substitution matrix (SM). Firstly, the theory parameter formulas relating to GOP, GAP, and SM are inferred, combining unaligned sequence length, number, and identity. Secondly, we tested the rationality of the theory parameter formulas, with experiment on the ClustalW and MAFFT program. In addition, we obtained a group of MAFFT program parameters according to the formulas proposed. The results of all experiments show that the SPS (sum-of-pair score) obtained from theory parameters is better than the SPS obtained from the default parameters of ClustalW and MAFFT. In both theory and practice, our method to determine the parameters is feasible and efficient. These can provide high-accuracy alignment results for precision medicine.},
 	note={29682519[pmid]},
 	note={PMC5842723[pmcid]},
@@ -77,13 +80,13 @@
 	and Davis, Trisha N.
 	and Riffle, Michael},
 	title={SnipViz: a compact and lightweight web site widget for display and dissemination of multiple versions of gene and protein sequences},
-	journal={BMC Res Notes},
+	journal={BMC Research Notes},
 	year={2014},
 	month={Jul},
 	day={23},
 	publisher={BioMed Central},
 	volume={7},
-	pages={468-468},
+	pages={468},
[…]
