### 2019-07-08 #1407 @csoneson
@IbrahimTanyalcin - we're almost there! I have sent a PR with some small edits to the paper, could you please take a look and merge this if you agree. Also, I noticed that for the Rauscher et al. and Wilm et al. references, the author list in the bibliography does not agree with the one on the journal website - can you fix this?
