### 2019-07-08 ffc1a12
update Rauscher et. al. and Wilm et. al. author fields
