+++ paper/paper.bib
@@ -36,7 +36,7 @@
 }
 
 @article{clustal,
-    author = {Wilm, A. and Higgins, D.G. and Valentin, F. and Blackshields, G. and McWilliam, H. and Wallace, I.M. and Thompson, J.D. and Larkin, M.A. and Brown, N.P. and McGettigan, P.A. and Chenna, R. and Lopez, R. and Gibson, T.J.},
+    author = {Larkin, M.A. and Blackshields, G. and Brown, N.P. and Chenna, R. and McGettigan, P.A. and McWilliam, H. and Valentin, F. and Wallace, I.M. and Wilm, A. and Lopez, R. and Thompson, J.D. and Gibson, T.J. and Higgins, D.G.},
     title = "{Clustal W and Clustal X version 2.0}",
     journal = {Bioinformatics},
     volume = {23},
@@ -97,7 +97,7 @@
 }
 
 @article{MSAViewer,
-    author = {Rauscher, Benedikt and Wilzbach, Sebastian and Goldberg, Tatyana and Rost, Burkhard and Yachdav, Guy and Sheridan, Robert and Sillitoe, Ian and Procter, James and Lewis, Suzanna E.},
+    author = {Yachdav, Guy and Wilzbach, Sebastian and Rauscher, Benedikt and Sheridan, Robert and Sillitoe, Ian and Procter, James and Lewis, Suzanna E. and Rost, Burkhard and Goldberg, Tatyana},
     title = "{MSAViewer: interactive JavaScript visualization of multiple sequence alignments}",
     journal = {Bioinformatics},
     volume = {32},
