### 2019-07-08 #1407 @IbrahimTanyalcin
:wave: @csoneson I rewrote the author fields, I've no idea why they were wrong in the first place. Thank you for pointing in out. I also merged your PR.
Thank you.
