### 2019-07-08 #1407 @IbrahimTanyalcin
@csoneson , before I even submitted for JOSS, I went ahead and created one:
https://zenodo.org/record/3272107#.XSOFeOszapo

I just released a new release now (ARIES), so the badge does not work yet, I guess it will work in a couple of hours. But it seems like it has tracked all the versions correctly. Is that ok?
