### 2019-07-08 #1407 @csoneson
Yes, the DOI of the latest release should be fine. However, the title and author list of the Zenodo archive must be the same as in your paper.
