### 2019-07-08 #1407 @IbrahimTanyalcin
@csoneson ok, I added [.zenodo.json](https://github.com/IbrahimTanyalcin/lexicon-mono-seq/blob/master/.zenodo.json), hopefully the webhook will work and it'll be picked up.
