### 2019-07-09 #1407 @IbrahimTanyalcin
:wave:  @csoneson the webhooks for zenodo works in a weird unpredictable way. I waited about half a day for things to be refreshed however I realized that won't happen.
So I went ahead and registered a new DOI manually. Is it better now?
