### 2019-07-09 #1407 @csoneson
@openjournals/joss-eics - we're ready to accept this paper!
